requirements.txt
streamlit
openai
python-docx
openpyxl
requests
beautifulsoup4
pymupdf
python-pptx